fix: video panel layout
